Untangle medium sized hash table from hash impl, add fibonacci hash
