+++ benches/benchmarks/mod.rs
@@ -2,7 +2,7 @@ pub mod fenwick;
 pub mod great_circle;
 pub mod k_way_merge_iterator;
 pub mod loser_tree;
+pub mod medium_size_hash_table;
 pub mod mercator;
 pub mod polyline;
 pub mod radix_sort;
-pub mod tabulation_hash;
