+++ src/lib.rs
@@ -19,6 +19,7 @@ pub mod edge;
 pub mod edmonds_karp;
 pub mod enumerative_source_coding;
 pub mod fenwick;
+pub mod fibonacci_hash;
 pub mod ford_fulkerson;
 pub mod geometry;
 pub mod graph;
@@ -34,6 +35,7 @@ pub mod loser_tree;
 pub mod lru;
 pub mod math;
 pub mod max_flow;
+pub mod medium_size_hash_table;
 pub mod mercator;
 pub mod merge_entry;
 pub mod merge_tree;
